# backend/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
    "/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "summary": "Register a new user",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User successfully registered"
          },
          "409": {
            "description": "User already exists"
          }
        },
        "tags": [
          "Authentication"
        ]
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "summary": "Login user",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User successfully logged in"
          },
          "401": {
            "description": "Invalid credentials"
          }
        },
        "tags": [
          "Authentication"
        ]
      }
    },
    "/auth/me": {
      "get": {
        "operationId": "AuthController_getProfile",
        "summary": "Get current user profile",
        "parameters": [],
        "responses": {
          "200": {
            "description": "User profile retrieved"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "tags": [
          "Authentication"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/clients": {
      "post": {
        "operationId": "ClientController_create",
        "summary": "Create a new client",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateClientDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Client successfully created"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "tags": [
          "Clients"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      },
      "get": {
        "operationId": "ClientController_findAll",
        "summary": "Get all clients for current user",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of clients"
          }
        },
        "tags": [
          "Clients"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/clients/{id}": {
      "get": {
        "operationId": "ClientController_findOne",
        "summary": "Get a specific client",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Client details"
          },
          "403": {
            "description": "Access denied"
          },
          "404": {
            "description": "Client not found"
          }
        },
        "tags": [
          "Clients"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      },
      "patch": {
        "operationId": "ClientController_update",
        "summary": "Update a client",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateClientDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Client successfully updated"
          },
          "403": {
            "description": "Access denied"
          },
          "404": {
            "description": "Client not found"
          }
        },
        "tags": [
          "Clients"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      },
      "delete": {
        "operationId": "ClientController_remove",
        "summary": "Delete a client",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Client successfully deleted"
          },
          "403": {
            "description": "Access denied"
          },
          "404": {
            "description": "Client not found"
          }
        },
        "tags": [
          "Clients"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/invoices": {
      "post": {
        "operationId": "InvoiceController_create",
        "summary": "Create a new invoice with line items",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateInvoiceDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Invoice successfully created with auto-calculated total"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Client not found or access denied"
          }
        },
        "tags": [
          "Invoices"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      },
      "get": {
        "operationId": "InvoiceController_findAll",
        "summary": "Get all invoices for current user",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of invoices with items and client info"
          }
        },
        "tags": [
          "Invoices"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/invoices/{id}": {
      "get": {
        "operationId": "InvoiceController_findOne",
        "summary": "Get a specific invoice",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Invoice details with items and client info"
          },
          "403": {
            "description": "Access denied"
          },
          "404": {
            "description": "Invoice not found"
          }
        },
        "tags": [
          "Invoices"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      },
      "patch": {
        "operationId": "InvoiceController_update",
        "summary": "Update an invoice",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateInvoiceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Invoice successfully updated with recalculated total"
          },
          "403": {
            "description": "Access denied"
          },
          "404": {
            "description": "Invoice not found"
          }
        },
        "tags": [
          "Invoices"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      },
      "delete": {
        "operationId": "InvoiceController_remove",
        "summary": "Delete an invoice",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Invoice successfully deleted"
          },
          "403": {
            "description": "Access denied"
          },
          "404": {
            "description": "Invoice not found"
          }
        },
        "tags": [
          "Invoices"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    }
  },
  "info": {
    "title": "Freelance CRM API",
    "description": "Production-ready API for managing freelance clients and invoices",
    "version": "1.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token",
        "in": "header"
      }
    },
    "schemas": {
      "RegisterDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "john.doe@example.com"
          },
          "password": {
            "type": "string",
            "example": "SecurePassword123!"
          },
          "name": {
            "type": "string",
            "example": "John Doe"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "john.doe@example.com"
          },
          "password": {
            "type": "string",
            "example": "SecurePassword123!"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "CreateClientDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Jane Smith"
          },
          "email": {
            "type": "string",
            "example": "jane.smith@example.com"
          },
          "company": {
            "type": "string",
            "example": "Acme Corporation"
          }
        },
        "required": [
          "name",
          "email"
        ]
      },
      "UpdateClientDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Jane Smith"
          },
          "email": {
            "type": "string",
            "example": "jane.smith@example.com"
          },
          "company": {
            "type": "string",
            "example": "Acme Corporation"
          }
        }
      },
      "CreateInvoiceItemDto": {
        "type": "object",
        "properties": {
          "description": {
            "type": "string",
            "example": "Website Development - Frontend"
          },
          "quantity": {
            "type": "number",
            "example": 40
          },
          "unitPrice": {
            "type": "number",
            "example": 75.5
          }
        },
        "required": [
          "description",
          "quantity",
          "unitPrice"
        ]
      },
      "CreateInvoiceDto": {
        "type": "object",
        "properties": {
          "invoiceNo": {
            "type": "string",
            "example": "INV-2024-001"
          },
          "status": {
            "type": "string",
            "enum": [
              "DRAFT",
              "PENDING",
              "PAID"
            ],
            "example": "DRAFT"
          },
          "dueDate": {
            "type": "string",
            "example": "2024-12-31T00:00:00.000Z"
          },
          "clientId": {
            "type": "string",
            "example": "client-uuid-here"
          },
          "items": {
            "example": [
              {
                "description": "Website Development",
                "quantity": 40,
                "unitPrice": 75.5
              },
              {
                "description": "Logo Design",
                "quantity": 5,
                "unitPrice": 100
              }
            ],
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateInvoiceItemDto"
            }
          }
        },
        "required": [
          "invoiceNo",
          "status",
          "dueDate",
          "clientId",
          "items"
        ]
      },
      "UpdateInvoiceDto": {
        "type": "object",
        "properties": {
          "invoiceNo": {
            "type": "string",
            "example": "INV-2024-001"
          },
          "status": {
            "type": "string",
            "enum": [
              "DRAFT",
              "PENDING",
              "PAID"
            ],
            "example": "DRAFT"
          },
          "dueDate": {
            "type": "string",
            "example": "2024-12-31T00:00:00.000Z"
          },
          "clientId": {
            "type": "string",
            "example": "client-uuid-here"
          },
          "items": {
            "example": [
              {
                "description": "Website Development",
                "quantity": 40,
                "unitPrice": 75.5
              },
              {
                "description": "Logo Design",
                "quantity": 5,
                "unitPrice": 100
              }
            ],
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateInvoiceItemDto"
            }
          }
        }
      }
    }
  }
}